pyproject.toml
[tool.poetry]
name = "markupsite"
version = "0.1.0"
description = ""
authors = ["Patrick Ryan <pmr3mh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
markdown-it-py = "^2.0.0"
mdit-py-plugins = "^0.3.0"
python-frontmatter = "^1.0.0"
beautifulsoup4 = "^4.10.0"
python-dateutil = "^2.8.2"
pytz = "^2021.3"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.9"
toml = "^0.10.2"
ipython = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
